Start documenting commands
Also experimenting with symlinking command docs as resources to the
application so the documentation can be used within the application. But
how do we want to render the reStructuredText? Pulling in sphinx as a
dep is a no-no (way too big and complex). Docutils would be fine,
though, if we can figure a way to render it nicely on the terminal?